Big merge fixes and ready to go !

+++ Vertical_Unity/Assets/Scripts/Enemies/OS_Bot.cs
@@ -17,7 +17,7 @@ public class OS_Bot : EnemyHandler
     public float snipAttackLockRange;
     public float lockingTime;
     public float chargingTime;
-    public float laserDamage;
+    public int laserDamage;
     public float laserBeamTime;
     public float laserStunTime;
     [Header("Debug settings")]
